requirements.txt
camoufox==0.4.11
gspread==6.1.2
google-auth==2.29.0
PyGithub==2.3.0
